Create STAR_GenomeIndexing.sh
This script automates the generation of a STAR aligner genome index for the hg38 human genome reference (see reference article used for the project by Hamelin et al., 2023). The script is tailored for high-performance computing (HPC) environments such as the university of Leicester's Alice HPC. It's configured for medium resource allocation with email notifications for job status updates for monitoring of the genome indexing process. It is suitable for subsequent large-scale RNA-seq analyses.