Create checksum_integrity.sh
This Bash script is made for use on HPC systems (e.g. University of Leicester's Alice HPC) managed with Slurm. It generates MD5 checksums for a large collection of FASTQ files, ensuring data integrity for downloaded fastq files. It is aimed at files prefixed with 'SRR' and ending in '.fastq', as we previously downloaded sequence read archive (SRA) data which was then converted to fastq files. The script automates the removal of any pre-existing checksum list to prevent data duplication, ensuring a fresh start for each run. The script includes email notifications set up for job start, end, and failure states to avoid the need of manual checks."