Create checksum_verification.sh
This script is designed for the automated verification of FASTQ file integrity using MD5 checksums on an HPC system. It submits a Slurm job that navigates to a scratch directory, and verifies the integrity of over 3000 FASTQ files against a  list of MD5 checksums obtained using the checksum_integrity.sh script. This process is crucial to ensure data integrity before proceeding with downstream analysis.