Add files via upload
This R script is designed to replicate the t-SNE visualizations presented in Figures 1E and 1D of the article. 

The purpose of the script is to:
1) Process the given data for dimensionality reduction via PCA.
2) Visualize the data in a two-dimensional t-SNE plot to observe clustering.
3) Highlight known cell types in the t-SNE plot as per the article's original figures.

Through these analyses, the script aims to validate the reproducibility of the published results and to demonstrate the clustering and differentiation of cell types within the dataset.